Split stable structures to files
